Trading

+++ Trading/Trading.Common/Entities/Price.cs
@@ -8,10 +8,10 @@ namespace Trading.Common.Entities
         public DateTime At { get; set; }
         public TimeSpan Span { get; set; }
 
-        public double Open { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Close { get; set; }
+        public decimal Open { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Close { get; set; }
 
         public long Volume { get; set; }
     }
